services/loader-api-candles/requirements.txt
aiohttp
uvloop
loguru
fastapi
python-dotenv
aiokafka